Fix caml_failwith vs failwith issue.
Using failwith generates a warning during build, and later a runtime
error.

Since I do not know whether this is due to me running ocaml >= 5.0,
and I do not plan to investigate at the moment, I simply added a
constraint to ocaml >= 5.0. This may be either an overkill, and maybe
completely irrelevant. This is not an issue for me here, since I do
plan on installing spin on switches with ocaml >= 5.0.